fix(core, daemon): address quality review on Phase 3
